pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pdf2podcast"
version = "0.1.2"
description = "A Python library to convert PDF documents into podcasts"
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT" }
authors = [
    { name = "Alberto Polini", email = "polini.alberto@gmail.com" }
]
dependencies = [
    "boto3>=1.34.0",           # For AWS Polly TTS
    "langchain-google-genai>=0.0.10",  # For Gemini integration
    "google-generativeai>=0.3.0",      # Required by langchain-google-genai
    "python-dotenv>=1.0.1",    # For environment variable management
    "PyMuPDF>=1.25.0",        # For advanced PDF processing
    "langchain-community>=0.0.21",  # Core LangChain functionality
    "pydantic>=2.6.0",        # For data validation
    "accelerate>=1.3.0",      # For ML model optimization
    "datasets>=3.2.0",        # For data handling
    "faiss-cpu>=1.10.0",      # For vector search
    "ffmpeg-python>=0.5.0",   # For audio processing
    "sentence-transformers>=3.4.1",  # For text embeddings
    "gTTS>=2.5.4",            # For Google Text-to-Speech
    "pydub>=0.25.1",          # For audio file manipulation (previously missing)
    "typing-extensions>=4.9.0"  # For advanced type hints
]

[project.urls]
Homepage = "https://github.com/albertopolini/pdf2podcast"
